fix: resolve all lint warnings and format issues
- Remove unused import of Bundle from hecs in engine.rs
- Remove unnecessary i32 casts in Position construction
- Fix field_reassign_with_default in test by using struct literal
- Apply cargo fmt to all source files for consistent formatting

All clippy warnings resolved, tests passing (12/12), code clean.

+++ crates/engine/src/engine.rs
@@ -2,9 +2,9 @@
 //!
 //! This module provides the deterministic simulation loop with entity component system.
 
-use hecs::{World, Bundle};
-use rand::SeedableRng;
+use hecs::World;
 use rand::Rng;
+use rand::SeedableRng;
 use rand_chacha::ChaCha8Rng;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
@@ -28,10 +28,10 @@ pub struct Position {
 /// Citizen entity component
 #[derive(Debug, Clone, Copy, PartialEq, Serialize, Deserialize)]
 pub struct Citizen {
-    pub age: u32,              // Age in years
-    pub health: Fixed,          // Health 0.0 - 1.0
-    pub ideology: Fixed,        // -1.0 (libertarian) to 1.0 (authoritarian)
-    pub welfare: Fixed,        // 0.0 - 1.0
+    pub age: u32,        // Age in years
+    pub health: Fixed,   // Health 0.0 - 1.0
+    pub ideology: Fixed, // -1.0 (libertarian) to 1.0 (authoritarian)
+    pub welfare: Fixed,  // 0.0 - 1.0
     pub job: Option<JobType>,
 }
 
@@ -72,14 +72,14 @@ pub struct Resources {
     pub food: Fixed,
     pub wood: Fixed,
     pub metal: Fixed,
-    pub energy: Fixed,  // Joules
+    pub energy: Fixed, // Joules
 }
 
 /// Production capability
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Production {
     pub output_type: ResourceType,
-    pub rate: Fixed,  // Per tick
+    pub rate: Fixed, // Per tick
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
@@ -158,23 +158,23 @@ impl Simulation {
     pub fn new() -> Self {
         let rng = SimRng::seed_from_u64(42);
         let mut world = World::new();
-        
+
         // Spawn initial entities
         Self::spawn_initial_entities(&mut world);
-        
+
         Self {
             state: WorldState::default(),
             world,
             rng,
         }
     }
-    
+
     /// Create simulation with custom seed
     pub fn with_seed(seed: u64) -> Self {
         let rng = SimRng::seed_from_u64(seed);
         let mut world = World::new();
         Self::spawn_initial_entities(&mut world);
-        
+
         Self {
             state: WorldState {
                 rng_seed: seed,
@@ -184,7 +184,7 @@ impl Simulation {
             rng,
         }
     }
-    
+
     /// Spawn initial world entities
     fn spawn_initial_entities(world: &mut World) {
         // Create initial citizens
@@ -198,7 +198,7 @@ impl Simulation {
             };
             let _ = world.spawn((citizen,));
         }
-        
+
         // Create city center
         let city = Building {
             building_type: BuildingType::CityCenter,
@@ -207,54 +207,54 @@ impl Simulation {
             position: Position { x: 0, y: 0 },
         };
         let _ = world.spawn((city,));
-        
+
         // Create farms
         for i in 0..5 {
             let farm = Building {
                 building_type: BuildingType::Farm,
                 hp: Fixed::from_num(200),
                 max_hp: Fixed::from_num(200),
-                position: Position { x: i as i32 - 2, y: 1 },
+                position: Position { x: i - 2, y: 1 },
             };
             let _ = world.spawn((farm,));
         }
-        
+
         // Create initial military
         for i in 0..10 {
             let soldier = MilitaryUnit {
                 unit_type: UnitType::Soldier,
                 strength: Fixed::from_num(10),
                 morale: Fixed::from_num(1),
-                position: Position { x: i as i32, y: 0 },
-                faction_id: 0,  // Player faction
+                position: Position { x: i, y: 0 },
+                faction_id: 0, // Player faction
             };
             let _ = world.spawn((soldier,));
         }
     }
-    
+
     /// Get mutable reference to RNG
     pub fn rng_mut(&mut self) -> &mut SimRng {
         &mut self.rng
     }
-    
+
     /// Advance simulation by one tick
     pub fn tick(&mut self) {
         self.state.tick += 1;
-        
+
         // Run simulation phases
         self.phase_production();
         self.phase_citizen_lifecycle();
         self.phase_military();
         self.phase_economy();
     }
-    
+
     /// Production phase - buildings produce resources
     fn phase_production(&mut self) {
         let mut production: HashMap<ResourceType, Fixed> = HashMap::new();
         production.insert(ResourceType::Food, Fixed::ZERO);
         production.insert(ResourceType::Wood, Fixed::ZERO);
         production.insert(ResourceType::Metal, Fixed::ZERO);
-        
+
         // Collect production from buildings
         for (_, building) in self.world.query::<&Building>().iter() {
             match building.building_type {
@@ -267,35 +267,37 @@ impl Simulation {
                 _ => {}
             }
         }
-        
+
         // Apply production to state (simplified - would go to resources in full impl)
-        tracing::debug!("Tick {} production: food={:?}, metal={:?}", 
+        tracing::debug!(
+            "Tick {} production: food={:?}, metal={:?}",
             self.state.tick,
             production.get(&ResourceType::Food),
-            production.get(&ResourceType::Metal));
+            production.get(&ResourceType::Metal)
+        );
     }
-    
+
     /// Citizen lifecycle phase
     fn phase_citizen_lifecycle(&mut self) {
         let mut births: u32 = 0;
-        
+
         for (_, citizen) in self.world.query::<&mut Citizen>().iter() {
             // Age citizens
             citizen.age += 1;
-            
+
             // Simple welfare decay/growth based on random
             let change = Fixed::from_num(self.rng.gen_range(-5..=5)) / Fixed::from_num(100);
             citizen.welfare = (citizen.welfare + change).clamp(Fixed::ZERO, Fixed::from_num(1));
         }
-        
+
         // Births based on welfare
         if self.state.population > 0 && self.rng.gen_bool(0.001) {
             births = 1;
         }
-        
+
         self.state.population += births as u64;
     }
-    
+
     /// Military phase
     fn phase_military(&mut self) {
         for (_, unit) in self.world.query::<&mut MilitaryUnit>().iter() {
@@ -306,21 +308,21 @@ impl Simulation {
             }
         }
     }
-    
+
     /// Economy phase - energy consumption
     fn phase_economy(&mut self) {
         // Base energy consumption per citizen
         let consumption = Fixed::from_num(self.state.population) / Fixed::from_num(1000);
-        self.state.energy_budget_joules = 
+        self.state.energy_budget_joules =
             (self.state.energy_budget_joules - consumption).max(Fixed::ZERO);
     }
-    
+
     /// Get snapshot of current state
     pub fn snapshot(&self) -> SimulationSnapshot {
         let citizen_count = self.world.query::<&Citizen>().iter().count();
         let building_count = self.world.query::<&Building>().iter().count();
         let military_count = self.world.query::<&MilitaryUnit>().iter().count();
-        
+
         SimulationSnapshot {
             tick: self.state.tick,
             population: self.state.population,
@@ -356,20 +358,20 @@ pub struct SimulationSnapshot {
 #[cfg(test)]
 mod tests {
     use super::*;
-    
+
     #[test]
     fn test_simulation_creation() {
         let sim = Simulation::new();
         assert_eq!(sim.state.tick, 0);
     }
-    
+
     #[test]
     fn test_tick_advances() {
         let mut sim = Simulation::new();
         sim.tick();
         assert_eq!(sim.state.tick, 1);
     }
-    
+
     #[test]
     fn test_initial_entities() {
         let sim = Simulation::new();
@@ -378,17 +380,17 @@ mod tests {
         assert!(snapshot.building_count > 0);
         assert!(snapshot.military_count > 0);
     }
-    
+
     #[test]
     fn test_determinism() {
         let mut sim1 = Simulation::with_seed(12345);
         let mut sim2 = Simulation::with_seed(12345);
-        
+
         for _ in 0..100 {
             sim1.tick();
             sim2.tick();
         }
-        
+
         assert_eq!(sim1.state.tick, sim2.state.tick);
         assert_eq!(sim1.state.population, sim2.state.population);
     }
